filesystem::path fixes

+++ include/inja/renderer.hpp
@@ -65,7 +65,7 @@ class Renderer : public NodeVisitor {
 
   static std::string htmlescape(const std::string& data) {
     std::string buffer;
-    buffer.reserve(1.1 * data.size());
+    buffer.reserve((unsigned int)(1.1 * data.size()));
     for (size_t pos = 0; pos != data.size(); ++pos) {
       switch (data[pos]) {
         case '&':  buffer.append("&amp;");       break;
